Extensive rewrite to implement loading of online, current weather information.  Added OpenWeather API but discovered that the most useful functions are only available for paying customers.  Implemented alternative Noaa API.  Added Image loader and download service.  Added Location service.  Added preview data.

+++ DemoWeather/Models/WeatherDay.swift
@@ -12,7 +12,7 @@ struct WeatherDay {
     let id = UUID()
     var day: Day = .MON
     var weather: Weather = Weather.Unknown
-    var temperature: Int = Int.min
+    var temperature: Double = Double(Int.min)
     
-    var temperatureText: String { get { temperature == Int.min ? "-°" : "\(temperature)°" } }
+    var temperatureText: String { get { temperature == Double(Int.min) ? "-°" : "\(temperature)°" } }
 }
